Implemented the first working wave Propagator.
Changed WaveSource so that it keeps track of source position.
Implemented the getDelayMatrix method in HexGrid.
Implemented the Propagator and SourcePropagator classes.
Fixed a bug in HeightGrid and ScaleGrid that would cause the grids to fly away :)
Changed default values for some of the HexGrids for better visuals.
Changed Hexagon colours to pink.